Added OcrManager
